% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrf1_testing.R
\name{nrf1_testing}
\alias{nrf1_testing}
\title{Net Random Forest Ver. 1 (NRF1) - Testing}
\usage{
nrf1_testing(X_testing, Y_testing, nrf1_mod, num_clusters, method)
}
\arguments{
\item{X_testing}{Features of the testing data.}

\item{Y_testing}{Supervisor of the testing data.}

\item{nrf1_mod}{A trained net random forest model.}

\item{num_clusters}{Number of clusters for the spectral clustering algorithm.}

\item{method}{"anova" (regression) or "class" (classification).}
}
\value{
A vector of predicted response values from the testing set.
}
\description{
Performs spectral clustering outside of all trees of the random forest on the testing data.
}
