% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrf2_testing.R
\name{nrf2_testing}
\alias{nrf2_testing}
\title{Net Random Forest Ver. 2 (NRF2) - Testing}
\usage{
nrf2_testing(X_testing, Y_testing, nrf2_mod, num_clusters, method)
}
\arguments{
\item{X_testing}{Features of the testing data.}

\item{Y_testing}{Supervisor of the testing data.}

\item{nrf2_mod}{A trained net random forest model.}

\item{num_clusters}{Number of clusters for the spectral clustering algorithm.}

\item{method}{"anova" (regression) or "class" (classification).}
}
\value{
A vector of predicted response values from the testing set.
}
\description{
Performs spectral clustering within each tree in the random forest on the testing data.
}
