% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrf2_training.R
\name{nrf2_training}
\alias{nrf2_training}
\title{Net Random Forest Ver. 2 (NRF2) - Training}
\usage{
nrf2_training(B, n, X_training, Y_training, num_clusters, prop_X = 0.8, method)
}
\arguments{
\item{B}{Number of trees.}

\item{n}{Number of subjects in each bootstrap sample.}

\item{X_training}{Features of the training data.}

\item{Y_training}{Supervisor of the training data.}

\item{num_clusters}{Number of clusters for the spectral clustering algorithm.}

\item{prop_X}{Proportion of features to be randomly sampled from the full set of features.}

\item{method}{"anova" (regression) or "class" (classification).}
}
\value{
A trained net random forest model that contains the predicted response values,
clustering memberships as well as a list of trees and selected features used in each tree.
}
\description{
Performs spectral clustering within each tree in the random forest on the training data.
}
